+++ src/Exception/ContainerException.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Petk\Container\Exception;
 
 use Psr\Container\ContainerExceptionInterface;