+++ src/Exception/EntryNotFoundException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Petk\Container\Exception;
-
-use Psr\Container\NotFoundExceptionInterface;
-
-class EntryNotFoundException extends \Exception implements NotFoundExceptionInterface
-{
-}