+++ tests/ContainerTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Petk\Container\Tests;
 
 use Petk\Container\Container;
+use Petk\Container\Exception\ContainerEntryNotFoundException;
 use Petk\Container\Exception\ContainerException;
 use Petk\Container\Tests\Fixtures\CircularReference\ChildClass;
 use Petk\Container\Tests\Fixtures\CircularReference\ParentClass;
@@ -13,6 +14,11 @@ use Petk\Container\Tests\Fixtures\Doer;
 use Petk\Container\Tests\Fixtures\Utility;
 use PHPUnit\Framework\Attributes\DataProvider;
 
+/**
+ * @internal
+ *
+ * @coversNothing
+ */
 class ContainerTest extends ContainerTestCase
 {
     #[DataProvider('keyProvider')]
@@ -30,6 +36,10 @@ class ContainerTest extends ContainerTestCase
 
         $container->set(Utility::class, function ($c) {
             return new Utility($c->get(Doer::class));
+        });
+
+        $container->set('foo-bar', function ($c) {
+            return new \stdClass();
         });
 
         $this->assertInstanceOf($valid, $container->get($key));
@@ -52,12 +62,37 @@ class ContainerTest extends ContainerTestCase
         $parent = $container->get(ParentClass::class);
     }
 
+    public function testSetInvalidEntry(): void
+    {
+        $this->expectException(ContainerException::class);
+
+        $container = new Container();
+
+        $container->set(ParentClass::class, new \stdClass());
+
+        $parent = $container->get(ParentClass::class);
+    }
+
+    public function testGetMissingEntry(): void
+    {
+        $this->expectException(ContainerEntryNotFoundException::class);
+
+        $container = new Container();
+
+        $container->set(\stdClass::class, function ($c) {
+            return new \stdClass();
+        });
+
+        $parent = $container->get(ParentClass::class);
+    }
+
     public static function keyProvider(): array
     {
         return [
             [Database::class, Database::class],
             [Doer::class, Doer::class],
             [Utility::class, Utility::class],
+            ['foo-bar', \stdClass::class],
         ];
     }
 }