+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] - YYYY-MM-DD
 
 ### Added
 
+- Fixed code style.
+- Exception names synced with the container context.
 - Added more tests and note about circular dependencies.
 
 ## [0.0.1] - 2023-06-05